requirements.txt
numpy
pandas
pytorch
torchvision
pytorch-cuda=11.7
tensorboard